Add super-easy script to regenerate AllTests.c
Premake doesn't let you add arbitrary actions and rules, so I can't
make AllTests.c automatically regen when the project is modified.
So as a workaround, it's now just a single short command to regen
them manually. I have the same problem with the sample dbs, so
premake may not work out in the long run.