Fix no-op save case
Of course when a card is saved with no modifications, its hash
doesn't change, which may confuse the datastore. So in the datastore,
it's now a no-op to save data with the same hash (before it would
delete the card completely!), and the schema uses the replace
conflict resolution policy on duplicate keys.