Got edges tracking movements of cards nicely
Except for cases where rise or run == 0. Main modifications are that
viewportcard geo_callback passes originating card, viewportedge stores
a list of coordinates, and there are separate get_coords and reset_coords,
the latter of which actually reads coords from card objects, and the former
of which just flattens self.coords.