fix geometry callback issues
I think because the first callback id issues is 0,
sometimes when setting an edge end for the first time in
the program it wouldn't realize it needed to remove callbacks.
Then when the original end was moved, it would call the callback
and the edge would be confused (raising that stupid string).